GODRIVER-3901 Move the List* logic to the mongo package. (#2501)
